+++ libs/my_c/src/string/my_wcstombs.c
@@ -17,12 +17,12 @@
  * @param dest The destination string
  * @param src The source string
  * @param n The number of bytes to convert
- * @return <b>intmax_t</b> The number of bytes converted if successful,
+ * @return <b>ssize_t</b> The number of bytes converted if successful,
  * the length of the string if dest is NULL, -1 if an error occured
  */
-intmax_t my_wcstombs(char *dest, const wchar_t *src, intmax_t n)
+ssize_t my_wcstombs(char *dest, const wchar_t *src, size_t n)
 {
-    intmax_t count = 0;
+    ssize_t count = 0;
 
     if (src == NULL || n == 0)
         return 0;
@@ -33,7 +33,7 @@ intmax_t my_wcstombs(char *dest, const wchar_t *src, intmax_t n)
     }
     if (dest == NULL)
         return count;
-    for (intmax_t index = 0; src[index] != L'\0' && index < n; index++)
+    for (size_t index = 0; src[index] != L'\0' && index < n; index++)
         dest[count] = (char)src[index];
     dest[count] = '\0';
     return n;
